updates via dashboard

+++ wp-content/plugins/newspack-newsletters/includes/service-providers/mailchimp/class-newspack-newsletters-mailchimp.php
@@ -1017,7 +1017,7 @@ final class Newspack_Newsletters_Mailchimp extends \Newspack_Newsletters_Service
 		}
 		if ( ! empty( $sender_email ) ) {
 			$is_valid_email = $this->validate_sender_email( $sender_email );
-			if ( is_wp_error( $is_valid_email ) ) {
+			if ( is_wp_error( $is_valid_email ) && $is_valid_email->get_error_code() === 'newspack_newsletters_unverified_sender_domain' ) {
 				delete_post_meta( $post->ID, 'senderEmail' ); // Delete invalid email so we can't accidentally attempt to send with it.
 				return $is_valid_email;
 			}
@@ -1414,26 +1414,21 @@ final class Newspack_Newsletters_Mailchimp extends \Newspack_Newsletters_Service
 	 * Throw an Exception if Mailchimp response indicates an error.
 	 *
 	 * @param Object $result Result of the Mailchimp operation.
-	 * @param String $preferred_error Preset error to use instead of Mailchimp errors.
+	 * @param String $preferred_error Error message to show to readers instead of showing Mailchimp API errors.
+	 * @param String $email_address Email address, for debugging purposes.
 	 * @throws Exception Error message.
 	 * @return The results of the API call.
 	 */
-	public function validate( $result, $preferred_error = null ) {
+	public function validate( $result, $preferred_error = null, $email_address = '' ) {
+		$default_error = __( 'Sorry, a Mailchimp error has occurred. Please try again later or contact us for support.', 'newspack-newsletters' );
+		if ( ! $preferred_error ) {
+			$preferred_error = $default_error;
+		}
 		if ( ! $result ) {
-			if ( $preferred_error ) {
-				throw new Exception( esc_html( $preferred_error ) );
-			} else {
-				throw new Exception( esc_html__( 'A Mailchimp error has occurred.', 'newspack-newsletters' ) );
-			}
+			throw new Exception( esc_html( $preferred_error ) );
 		}
 		// See Mailchimp error code glossary: https://mailchimp.com/developer/marketing/docs/errors/#error-glossary.
 		if ( ! empty( $result['status'] ) && (int) $result['status'] >= 400 ) {
-			if ( $preferred_error && ! Newspack_Newsletters::debug_mode() ) {
-				if ( ! empty( $result['detail'] ) ) {
-					$preferred_error .= ' ' . $result['detail'];
-				}
-				throw new Exception( esc_html( $preferred_error ) );
-			}
 			$messages = [];
 			if ( ! empty( $result['errors'] ) ) {
 				foreach ( $result['errors'] as $error ) {
@@ -1442,13 +1437,36 @@ final class Newspack_Newsletters_Mailchimp extends \Newspack_Newsletters_Service
 					}
 				}
 			}
-			if ( ! count( $messages ) && ! empty( $result['detail'] ) ) {
+			if ( ! empty( $result['detail'] ) ) {
 				$messages[] = $result['detail'];
 			}
 			if ( ! count( $messages ) ) {
-				$message[] = __( 'A Mailchimp error has occurred.', 'newspack-newsletters' );
+				$message[] = $preferred_error;
 			}
-			throw new Exception( esc_html( implode( ' ', $messages ) ) );
+
+			// Log the error with any details returned from the API.
+			do_action(
+				'newspack_log',
+				'newspack_mailchimp_api_error',
+				'Error adding contact to Mailchimp.',
+				[
+					'type'       => 'error',
+					'data'       => [
+						'messages' => $messages,
+						'status'   => $result['status'],
+						'title'    => ! empty( $result['title'] ) ? $result['title'] : '',
+					],
+					'user_email' => $email_address,
+					'file'       => 'newspack_mailchimp',
+				]
+			);
+
+			if ( Newspack_Newsletters::debug_mode() ) {
+				// Show the "real" error message(s) if in debug mode.
+				throw new Exception( esc_html( implode( ' ', $messages ) ) );
+			} else {
+				throw new Exception( esc_html( $preferred_error ) );
+			}
 		}
 		return $result;
 	}
@@ -1814,8 +1832,16 @@ final class Newspack_Newsletters_Mailchimp extends \Newspack_Newsletters_Service
 			Newspack_Newsletters_Logger::log( 'Mailchimp add_contact PUT payload: ' . wp_json_encode( $update_payload ) );
 
 			// Create or update a list member.
-			$member_hash = Mailchimp::subscriberHash( $email_address );
-			$result      = $this->validate( $mc->put( "lists/$list_id/members/$member_hash", $update_payload ) );
+			$member_hash  = Mailchimp::subscriberHash( $email_address );
+			$reader_error = $this->get_add_contact_reader_error_message(
+				[
+					'email'     => $contact['email'],
+					'list_id'   => $list_id,
+					'tags'      => $tags,
+					'interests' => $interests,
+				]
+			);
+			$result       = $this->validate( $mc->put( "lists/$list_id/members/$member_hash", $update_payload ), $reader_error, $email_address );
 		} catch ( \Exception $e ) {
 			return new \WP_Error(
 				'newspack_newsletters_mailchimp_add_contact_failed',
